requirements.txt
openai>=1.0.0
gemini-webapi>=0.3.0
python-dotenv>=1.0.0
redis>=5.0.0
httpx>=0.27.0
Pillow>=10.0.0
